perf(map): cache NR inventory for /api/map/rentals + cron warm
Use unstable_cache for full listing pagination; bbox filtering stays in-memory.
Add secured cron route and Vercel schedule to warm cache after deploys.

Made-with: Cursor
