+++ vercel.json
@@ -1,0 +1,8 @@
+{
+  "crons": [
+    {
+      "path": "/api/cron/warm-map-rentals",
+      "schedule": "*/10 * * * *"
+    }
+  ]
+}
